requirements.txt
decord==0.5.2
imageio-ffmpeg==0.4.3
tensorflow==2.4.1
tqdm==4.57.0
wandb==0.10.18
yacs==0.1.8
